+++ src/main/java/lotto/domain/LottoNumber.java
@@ -1,22 +1,15 @@
-package lotto;
+package lotto.domain;
 
-import java.util.Comparator;
 import java.util.List;
 
-public class Lotto {
+public class LottoNumber {
     private static final int FIRST_INDEX = 0;
     private static final int LAST_INDEX = 5;
     private static final int MIN_LOTTO_NUMBER = 1;
     private static final int MAX_LOTTO_NUMBER = 45;
 
-    private static final String NOT_DISTINCT_NUMBERS_ERROR_MSG = "중복되는 로또 번호가 존재합니다.";
     private static final String NOT_IN_SCOPE_NUMBERS_ERROR_MSG = "로또 번호의 범위가 잘못되었습니다.";
-
-    public static void validateDistinctNumbers(List<Integer> inputNumbers) {
-        if (inputNumbers.stream().distinct().count() != inputNumbers.size()) {
-            throw new IllegalArgumentException(NOT_DISTINCT_NUMBERS_ERROR_MSG);
-        }
-    }
+    private static final String NOT_NUMBER_MSG = "정수로 입력하셔야 합니다.";
 
     public static void validateNumberScope(List<Integer> invalidNumbersUnderScope) {
         invalidNumbersUnderScope.sort(Integer::compareTo);
@@ -25,4 +18,12 @@ public class Lotto {
             throw new IllegalArgumentException(NOT_IN_SCOPE_NUMBERS_ERROR_MSG);
         }
     }
+    public static void validateNumber(String invalidInputMoney) {
+        try{
+            Integer.parseInt(invalidInputMoney);
+        } catch (NumberFormatException e){
+            throw new IllegalArgumentException(NOT_NUMBER_MSG);
+        }
+    }
 }
+
