+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,3 +1,6 @@
+package lotto.domain;
+
+import lotto.domain.Money;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -19,13 +22,5 @@ public class MoneyTest {
         assertThatThrownBy(() -> Money.validateUnderLottoPrice(invalidInputMoney))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("1000원 이상으로 입력해야 합니다.");
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"천원", "3.2"})
-    void validateNumber_로또_가격을_정수로_입력하지_않았을_때(String invalidInputMoney) {
-        assertThatThrownBy(() -> Money.validateNumber(invalidInputMoney))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("정수로 입력하셔야 합니다.");
     }
 }
