+++ src/test/java/lotto/domain/NumberTest.java
@@ -1,0 +1,16 @@
+package lotto.domain;
+
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+public class NumberTest {
+    @ParameterizedTest
+    @ValueSource(strings = {"천원", "3.2"})
+    void validateNumber_정수로_입력하지_않았을_때(String invalidInputMoney) {
+        assertThatThrownBy(() -> Number.validateNumber(invalidInputMoney))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("정수로 입력하셔야 합니다.");
+    }
+}
