+++ src/test/java/lotto/domain/WinningLottoTest.java
@@ -1,0 +1,14 @@
+package lotto.domain;
+
+public class WinningLottoTest {
+//    @ParameterizedTest
+//    @ValueSource(strings = {"천원", "3.2"})
+//    void validateNumber_로또_가격을_정수로_입력하지_않았을_때(String invalidInputMoney) {
+//        assertThatThrownBy(() -> WinningLotto.validateNumber(invalidInputMoney))
+//                .isInstanceOf(IllegalArgumentException.class)
+//                .hasMessage("정수로 입력하셔야 합니다.");
+//    }
+
+
+
+}
